+++ packages/client/next.config.js
@@ -1,7 +1,19 @@
+const path = require('path')
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-  styledComponents: true
+  trailingSlash: false,
+  webpackDevMiddleware: config => {
+    config.watchOptions = {
+      poll: 1000,
+      aggregateTimeout: 300
+    }
+
+    return config
+  },
+  sassOptions: {
+    includePaths: [path.join(__dirname, 'styles')]
+  }
 }
 
 module.exports = nextConfig
