+++ packages/client/postcss.config.js
@@ -1,0 +1,17 @@
+module.exports = {
+  plugins: {
+    'postcss-import': {},
+    'autoprefixer': {},
+    'tailwindcss': {},
+    'postcss-flexbugs-fixes': {},
+    'postcss-preset-env': {
+      autoprefixer: {
+        flexbox: 'no-2009'
+      },
+      stage: 3,
+      features: {
+        'custom-properties': false
+      }
+    }
+  }
+}
